requirements.txt
beautifulsoup4==4.14.3
fastapi==0.128.7
langchain==1.2.10
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.10
langchain_ollama==1.0.1
langchain_text_splitters==1.1.0
langgraph==1.0.8
langserve==0.3.3
pooch==1.8.2
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0
